+++ mayan/apps/document_states/serializers.py
@@ -70,9 +70,10 @@ class WorkflowTemplateSerializer(serializers.HyperlinkedModelSerializer):
             }
         }
         fields = (
-            'document_types_add_url', 'document_types_remove_url',
-            'document_types_url', 'id', 'image_url', 'internal_name',
-            'label', 'states_url', 'transitions_url', 'url'
+            'auto_launch', 'document_types_add_url',
+            'document_types_remove_url', 'document_types_url', 'id',
+            'image_url', 'internal_name', 'label', 'states_url',
+            'transitions_url', 'url'
         )
         model = Workflow
         read_only_fields = (